separate cursor and question
