+++ Assets/Scripts/Text/Question.cs
@@ -10,25 +10,19 @@ public class Question : MonoBehaviour
     [SerializeField] private GameObject[] questionBranches;
     private Image[] questionBranchImage;
 
-    private InputSetting _inputSetting;
-    private int cursorPlace;
-    [SerializeField] private Sprite cursorSprite;  //カーソル画像
-    [SerializeField] private Sprite questionBranchSprite;  //非選択時の画像
-
-    private int cursorMax;
     string[][] word;
-    public bool thinkingTime;
+    private bool thinkingTime;
+    private int cursorMax;
 
     void Start()
     {
-        _inputSetting = InputSetting.Load();
+        
         questionBranchImage = new Image[questionBranches.Length];
         for (int i = 0; i < questionBranches.Length; i++)
         {
             questionBranches[i].SetActive(false);
             questionBranchImage[i] = questionBranches[i].GetComponent<Image>();
         }
-        cursorPlace = 0;
         thinkingTime = false;
     }
 
@@ -39,6 +33,15 @@ public class Question : MonoBehaviour
         QuestionBranch(word[0]);
         cursorMax = Mathf.Min(word[0].Length, questionBranches.Length);
     }
+    public int GetCursorMax()
+    {
+        return cursorMax;
+    }
+
+    public bool GetThinkingTime()
+    {
+        return thinkingTime;
+    }
 
     public void DisableQuestionBranches()
     {
@@ -47,6 +50,11 @@ public class Question : MonoBehaviour
             questionBranches[i].SetActive(false);
             thinkingTime = false;
         }
+    }
+
+    public Image[] GetQuestionBranchImage() 
+    { 
+        return questionBranchImage; 
     }
 
     private void QuestionBranch(string[] str)
@@ -60,32 +68,7 @@ public class Question : MonoBehaviour
         thinkingTime = true;
     }
 
-    public int CursorMax()
-    {
-        return cursorMax;
-    }
-
-    public void CursorMove(int max)
-    {
-        if (cursorPlace >= 0 && cursorPlace < max) 
-        {
-            if (_inputSetting.GetBackKeyUp() || _inputSetting.GetRightKeyUp())  //一直線に並べることを想定
-            {
-                cursorPlace = Mathf.Min(cursorPlace + 1, max - 1);
-            }
-            else if (_inputSetting.GetForwardKeyUp() || _inputSetting.GetLeftKeyUp())
-            {
-                cursorPlace = Mathf.Max(cursorPlace - 1, 0);
-            }
-        }
-        for (int i = 0; i < max; ++i)
-        {
-            questionBranchImage[i].sprite = questionBranchSprite;
-        }
-        questionBranchImage[cursorPlace].sprite = cursorSprite;
-    }
-
-    public void QuestionOutput()
+    public void QuestionOutput(int cursorPlace)
     {
         DebugLogger.Log(word[1][cursorPlace]);  //仮の出力
     }
