+++ Assets/Scripts/Text/TextWindowCursor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5309f632caa58ba4ab190f0f7f34fc79
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
